feat: Introduce Android layer services model & refactoring
