+++ android/src/main/kotlin/com/wultra/android/powerauth/flutter/internal/core/PowerAuthFlutterService.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2025 Wultra s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.wultra.android.powerauth.flutter.internal.core
+
+import io.flutter.plugin.common.MethodCall
+import io.flutter.plugin.common.MethodChannel.Result
+
+sealed interface PowerAuthFlutterService {
+    val name: String
+    val handlers: Map<String, MethodHandler>
+
+    fun interface MethodHandler {
+        fun handle(call: MethodCall, result: Result)
+    }
+}
